Move Equatable operators into its own file.
